# perf: use Object instead of Map for promise table