This PR addresses GC problems with `deno_tcp.ts`

Benchmark:
```
$ deno --allow-net tools/deno_tcp.ts
$ wrk -d 10 --latency http://127.0.0.1:4500
```

this PR
```
Running 10s test @ http://127.0.0.1:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   135.29us   41.03us   3.14ms   97.11%
    Req/Sec    36.23k     1.34k   37.46k    96.53%
  Latency Distribution
     50%  131.00us
     75%  133.00us
     90%  137.00us
     99%  287.00us
  727913 requests in 10.10s, 35.40MB read
Requests/sec:  72072.56
Transfer/sec:      3.51MB
```

`master`
```
Running 10s test @ http://127.0.0.1:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   197.70us  398.18us   5.68ms   96.89%
    Req/Sec    34.59k     2.38k   38.25k    85.15%
  Latency Distribution
     50%  128.00us
     75%  134.00us
     90%  166.00us
     99%    2.70ms
  695113 requests in 10.10s, 33.81MB read
Requests/sec:  68814.49
Transfer/sec:      3.35MB
```

`v0.35.0`
```
Running 10s test @ http://127.0.0.1:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   285.51us  711.19us   9.51ms   95.88%
    Req/Sec    31.45k     2.56k   35.05k    88.00%
  Latency Distribution
     50%  139.00us
     75%  144.00us
     90%  189.00us
     99%    4.60ms
  625922 requests in 10.00s, 30.44MB read
Requests/sec:  62572.10
Transfer/sec:      3.04MB
```

GC output
this PR
```
[72564:0x38b300000000]     8904 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     8920 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     8940 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     8960 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     8978 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.1 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     8995 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.1 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9011 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9038 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9054 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9070 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9086 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.2 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9102 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.1 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9119 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.4 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9135 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.1 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9151 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.1 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
[72564:0x38b300000000]     9168 ms: Scavenge 3.4 (6.2) -> 1.4 (6.2) MB, 0.4 / 0.0 ms  (average mu = 1.000, current mu = 1.000) allocation failure
```

`master`
```
[82117:0x13a300000000]     8862 ms: Scavenge 36.3 (53.0) -> 24.1 (55.5) MB, 3.4 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     8973 ms: Scavenge 39.3 (56.2) -> 27.3 (59.0) MB, 3.5 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     8996 ms: Mark-sweep 30.2 (59.0) -> 1.5 (36.5) MB, 1.0 / 0.0 ms  (+ 0.4 ms in 3 steps since start of marking, biggest step 0.2 ms, walltime since start of marking 3 ms) (average mu = 0.999, current mu = 0.999) finalize incremental marking via stack guard GC in old space requested
[82117:0x13a300000000]     9129 ms: Scavenge 18.4 (36.5) -> 5.3 (36.5) MB, 2.8 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9237 ms: Scavenge 19.7 (36.5) -> 8.2 (40.2) MB, 4.7 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9354 ms: Scavenge 23.5 (40.2) -> 11.5 (43.2) MB, 3.3 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9470 ms: Scavenge 26.6 (43.5) -> 14.6 (46.5) MB, 3.5 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9583 ms: Scavenge 29.8 (46.5) -> 17.7 (49.5) MB, 3.6 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9696 ms: Scavenge 33.0 (49.7) -> 20.9 (53.0) MB, 3.7 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9809 ms: Scavenge 36.2 (53.0) -> 24.1 (55.7) MB, 3.4 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9920 ms: Scavenge 39.4 (56.2) -> 27.4 (59.0) MB, 3.6 / 0.0 ms  (average mu = 0.999, current mu = 0.999) allocation failure
[82117:0x13a300000000]     9939 ms: Mark-sweep 29.8 (59.0) -> 1.4 (36.7) MB, 0.8 / 0.0 ms  (+ 0.2 ms in 2 steps since start of marking, biggest step 0.1 ms, walltime since start of marking 2 ms) (average mu = 0.999, current mu = 0.999) finalize incremental marking via stack guard GC in old space requested
```