@YurySolovyov 2020-03-10
Should this be reported to V8?