@ry 2020-03-11
This PR indeed improved latency for deno_tcp

<img width="947" alt="Screen Shot 2020-03-11 at 11 26 22 AM" src="https://user-images.githubusercontent.com/80/76433932-64256e00-638b-11ea-9962-15b66ee4aa99.png">

The benefit for throughput is marginally better

<img width="957" alt="Screen Shot 2020-03-11 at 11 28 26 AM" src="https://user-images.githubusercontent.com/80/76434123-a3ec5580-638b-11ea-9aa6-7bb1d4e40ad2.png">